# movie_express/docs/swagger.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Movie information API
  description: This API provides access to publicly available data on a limited number of movies published from the year 1990 from the Internet Movie Database and other sources. The API endpoints and their usage are described in detail below.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Enter JWT Bearer token **_only_**
  schemas:
...
    MissingAuthHeader:
      type: object
      properties:
        error:
          type: boolean
          example: true
        message:
          type: string
          example: Authorization header ('Bearer token') not found
    TokenExpired:
      type: object
      properties:
        error:
          type: boolean
          example: true
        message:
          type: string
          example: JWT token has expired
    InvalidJWT:
      type: object
      properties:
        error:
          type: boolean
          example: true
        message:
          type: string
          example: Invalid JWT token
...
    InvalidFieldTypes:
      type: object
      properties:
        error:
          type: boolean
          example: true
        message:
          type: string
          example: "Request body invalid: firstName, lastName, dob, and address must be strings only"
    InvalidDateFormat:
      type: object
      properties:
        error:
          type: boolean
          example: true
        message:
          type: string
          example: "Invalid input: dob must be a real date in format YYYY-MM-DD"
paths:
...
  /user/register:
    post:
      tags:
        - Authentication
      description: Creates a new user account. A request body containing the user to be registered must be sent.
      requestBody:
        description: An object containing the email and password of the user to be registered.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: mike@gmail.com
                password:
                  type: string
                  example: password
      responses:
        201:
          description: User successfully created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User created
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Request body incomplete, both email and password are required
        409:
          description: User already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: User already exists
        429:
          description: Rate limit exceeded.
          content:
            text/html:
              schema:
                type: string
                example: Too many requests, please try again later.
  /user/login:
    post:
      description: Log in to an existing user account. A request body containing the user credentials must be sent. The longExpiry bool is a setting for development use only that makes both tokens expire after a year.
      tags:
        - Authentication
      requestBody:
        description: The credentials of the user to log in.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: mike@gmail.com
                password:
                  type: string
                  example: password
                longExpiry:
                  type: boolean
                  example: false
                bearerExpiresInSeconds:
                  type: number
                  example: 600
                refreshExpiresInSeconds:
                  type: boolean
                  example: 86400
      responses:
        200:
          description: Log in successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  bearerToken:
                    type: object
                    properties:
                      token:
                        type: string
                        example: ajsonwebtoken
                      token_type:
                        type: string
                        example: Bearer
                      expires_in:
                        type: number
                        example: 600
                  refreshToken:
                    type: object
                    properties:
                      token:
                        type: string
                        example: ajsonwebtoken
                      token_type:
                        type: string
                        example: Refresh
                      expires_in:
                        type: number
                        example: 86400
        400:
          description: Invalid log in request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Request body incomplete, both email and password are required
        401:
          description: Log in failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Incorrect email or password
        429:
          description: Rate limit exceeded.
          content:
            text/html:
              schema:
                type: string
                example: Too many requests, please try again later.
  /user/refresh:
    post:
      description: Obtain a new bearer token by using a refresh token
      tags:
        - Authentication
      requestBody:
        description: The refresh token
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  example: ajsonwebtoken
      responses:
        200:
          description: Token successfully refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  bearerToken:
                    type: object
                    properties:
                      token:
                        type: string
                        example: ajsonwebtoken
                      token_type:
                        type: string
                        example: Bearer
                      expires_in:
                        type: number
                        example: 600
                  refreshToken:
                    type: object
                    properties:
                      token:
                        type: string
                        example: ajsonwebtoken
                      token_type:
                        type: string
                        example: Refresh
                      expires_in:
                        type: number
                        example: 86400
        400:
          description: Invalid refresh request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Request body incomplete, refresh token required
        401:
          description: Unauthorized. Click on 'Schema' below to see the possible error responses.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenExpired"
        429:
          description: Rate limit exceeded.
          content:
            text/html:
              schema:
                type: string
                example: Too many requests, please try again later.
  /user/logout:
    post:
      description: Log the user out, invalidating the refresh token
      tags:
        - Authentication
      requestBody:
        description: The refresh token
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  example: ajsonwebtoken
      responses:
        200:
          description: Token successfully invalidated
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Token successfully invalidated
        400:
          description: Invalid refresh request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Request body incomplete, refresh token required
        401:
          description: Unauthorized. Click on 'Schema' below to see the possible error responses.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenExpired"
        429:
          description: Rate limit exceeded.
          content:
            text/html:
              schema:
                type: string
                example: Too many requests, please try again later.
  /user/{email}/profile:
    get:
      description: Returns a user’s profile information as a JSON object.
      tags:
        - Profile
      parameters:
        - name: email
          in: path
          description: User's email address.
          required: true
          schema:
            type: string
      responses:
        200:
          description: Successful request. Returns the user's profile information as a JSON object.
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                    example: "mike@gmail.com"
                  firstName:
                    type: string
                    example: "Michael"
                  lastName:
                    type: string
                    example: "Jordan"
                  dob:
                    type: string
                    example: "1963-02-17"
                  address:
                    type: string
                    example: "123 Fake Street, Springfield"
              examples:
                Unauthorized:
                  value:
                    email: "mike@gmail.com"
                    firstName: "Michael"
                    lastName: "Jordan"
                Authorized:
                  value:
                    email: "mike@gmail.com"
                    firstName: "Michael"
                    lastName: "Jordan"
                    dob: "1963-02-17"
                    address: "123 Fake Street, Springfield"
                New_user:
                  value:
                    email: "notmike@gmail.com"
                    firstName: null
                    lastName: null
                    dob: null
                    address: null

        401:
          description: Unauthorized. Click on 'Schema' below to see the possible error responses.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Authorization header is malformed
                oneOf:
                  - $ref: "#/components/schemas/TokenExpired"
                  - $ref: "#/components/schemas/InvalidJWT"
        404:
          description: Not Found. There is no user corresponds to email.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: User not found
    put:
      summary: Updates a user's profile information.
      tags:
        - Profile
      parameters:
        - name: email
          in: path
          description: User's email address.
          required: true
          schema:
            type: string

      responses:
        200:
          description: Profile successfully updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                    example: "mike@gmail.com"
                  firstName:
                    type: string
                    example: "Michael"
                  lastName:
                    type: string
                    example: "Jordan"
                  dob:
                    type: string
                    format: date
                    example: "1963-02-17"
                  address:
                    type: string
                    example: "123 Fake Street, Springfield"
        400:
          description: Bad Request. The submitted object is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Request body incomplete: firstName, lastName, dob, and address are required"
                oneOf:
                  - $ref: "#/components/schemas/InvalidFieldTypes"
                  - $ref: "#/components/schemas/InvalidDateFormat"

        401:
          description: Unauthorized. The user is not authorized to update the profile.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Authorization header is malformed
                oneOf:
                  - $ref: "#/components/schemas/MissingAuthHeader"
                  - $ref: "#/components/schemas/TokenExpired"
                  - $ref: "#/components/schemas/InvalidJWT"
        403:
          description: Forbidden. The user is attempting to update another user's profile.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Forbidden"
        404:
          description: Not Found. There is no user corresponding to the provided email.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "User not found"